adding dtype=np.int64 to all np.sum() calls

+++ porespy/generators/_imgen.py
@@ -1178,7 +1178,7 @@ def _cylinders(
     elif np.size(shape) == 2:
         raise Exception("2D cylinders don't make sense")
     # Find hypotenuse of domain from [0,0,0] to [Nx,Ny,Nz]
-    H = np.sqrt(np.sum(np.square(shape))).astype(int)
+    H = np.sqrt(np.sum(np.square(shape)), dtype=np.int64).astype(int)
     if length is None:  # Assume cylinders span domain if length not given
         length = 2 * H
     R = min(int(length / 2), 2 * H)  # Trim given length to 2H if too long
@@ -1350,7 +1350,7 @@ def cylinders(
     # Rough estimate of n_fibers
     n_fibers_added = 0
     # Calculate fraction of fibers to be added in each iteration.
-    subdif = 0.8 / np.sum(np.arange(1, maxiter) ** 2)
+    subdif = 0.8 / np.sum(np.arange(1, maxiter) ** 2, dtype=np.int64)
     fractions = [0.2]
     for i in range(1, maxiter):
         fractions.append(fractions[i - 1] + (maxiter - i) ** 2 * subdif)
